requirements.txt
torch
qdrant-client
langchain-community
langchain-qdrant
langchain-text-splitters
sentence-transformers
tiktoken
PyPDF2
python-docx
python-pptx
uvicorn
fastapi
openai
accelerate
streamlit
